+++ LightCTR/train/train_ffm_algo.cpp
@@ -14,7 +14,7 @@ void Train_FFM_Algo::init() {
     
     learnable_params_cnt = this->feature_cnt * this->field_cnt * this->factor_cnt
                            + this->feature_cnt;
-    update_g = new double[learnable_params_cnt];
+    update_g = new float[learnable_params_cnt];
     updater.learnable_params_cnt(learnable_params_cnt);
     
     printf("Training FFM\n");
@@ -30,7 +30,7 @@ void Train_FFM_Algo::Train() {
         
         for (size_t p = 0; p < minibatch_epoch; p++) {
             
-            memset(update_g, 0, sizeof(double) * learnable_params_cnt);
+            memset(update_g, 0, sizeof(float) * learnable_params_cnt);
             
             size_t start_pos = p * batch_size;
             batchGradCompute(start_pos, min(start_pos + batch_size, this->dataRow_cnt));
@@ -45,12 +45,12 @@ void Train_FFM_Algo::Train() {
 void Train_FFM_Algo::batchGradCompute(size_t rbegin, size_t rend) {
     for (size_t rid = rbegin; rid < rend; rid++) { // data row
         threadpool->addTask([&, rid]() {
-            double fm_pred = 0.0f, fm_pred_field = 0.0f;
+            float fm_pred = 0.0f, fm_pred_field = 0.0f;
             
             for (size_t i = 0; i < dataSet[rid].size(); i++) {
                 const size_t fid = dataSet[rid][i].first;
                 assert(fid < this->feature_cnt);
-                const double X = dataSet[rid][i].second;
+                const float X = dataSet[rid][i].second;
                 const size_t field = dataSet[rid][i].field;
                 
                 fm_pred += W[fid] * X;
@@ -60,13 +60,13 @@ void Train_FFM_Algo::batchGradCompute(size_t rbegin, size_t rend) {
                         continue;
                     }
                     const size_t fid2 = dataSet[rid][j].first;
-                    const double X2 = dataSet[rid][j].second;
+                    const float X2 = dataSet[rid][j].second;
                     const size_t field2 = dataSet[rid][j].field;
                     
-                    double field_w = 0;
+                    float field_w = 0;
                     for (size_t fac_itr = 0; fac_itr < this->factor_cnt; fac_itr++) {
-                        double v1 = *getV_field(fid, field2, fac_itr);
-                        double v2 = *getV_field(fid2, field, fac_itr);
+                        float v1 = *getV_field(fid, field2, fac_itr);
+                        float v2 = *getV_field(fid2, field, fac_itr);
                         field_w += v1 * v2;
                     }
                     fm_pred_field += field_w * X * X2;
@@ -78,8 +78,8 @@ void Train_FFM_Algo::batchGradCompute(size_t rbegin, size_t rend) {
     threadpool->wait();
 }
 
-void Train_FFM_Algo::accumWVGrad(size_t rid, double pred) {
-    const double loss = pred - label[rid];
+void Train_FFM_Algo::accumWVGrad(size_t rid, float pred) {
+    const float loss = pred - label[rid];
     if (loss == 0) {
         return;
     }
@@ -108,8 +108,8 @@ void Train_FFM_Algo::accumWVGrad(size_t rid, double pred) {
             {
                 unique_lock<SpinLock> glock(this->lock_v);
                 for (size_t fac_itr = 0; fac_itr < this->factor_cnt; fac_itr++) {
-                    const double v1 = *getV_field(fid, field2, fac_itr);
-                    const double v2 = *getV_field(fid2, field, fac_itr);
+                    const float v1 = *getV_field(fid, field2, fac_itr);
+                    const float v2 = *getV_field(fid2, field, fac_itr);
                     *update_V(fid, field2, fac_itr) += loss * x2 * v2 + L2Reg_ratio * v1;
                     *update_V(fid2, field, fac_itr) += loss * x2 * v1 + L2Reg_ratio * v2;
                 }
@@ -121,7 +121,7 @@ void Train_FFM_Algo::accumWVGrad(size_t rid, double pred) {
 void Train_FFM_Algo::ApplyGrad() {
     updater.update(0, this->feature_cnt, W, update_g);
     
-    double *gradV = update_g + this->feature_cnt;
+    float *gradV = update_g + this->feature_cnt;
     updater.update(this->feature_cnt, this->feature_cnt *
                    this->field_cnt * this->factor_cnt, V, gradV);
 }
